requirements.txt
numpy
scipy
Pillow
matplotlib
scikit-learn
scikit-image
opencv-python
pandas
xgboost
tqdm
jupyter